task/requirements.txt
Django==4.2.6
psycopg2-binary==2.9.6
djangorestframework==3.14.0
rest_framework_simplejwt==5.1.0
drf-yasg==1.21.4